updated t_vector to float

+++ include/cub3d.h
@@ -6,7 +6,7 @@
 /*   By: abasdere <abasdere@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/04 11:15:26 by averin            #+#    #+#             */
-/*   Updated: 2024/03/07 13:02:05 by abasdere         ###   ########.fr       */
+/*   Updated: 2024/03/11 10:54:01 by abasdere         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,8 +22,8 @@
 
 typedef struct s_vector
 {
-	double	x;
-	double	y;
+	float	x;
+	float	y;
 }	t_vector;
 
 typedef struct s_player
